% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_frequency.R
\name{app_frequency}
\alias{app_frequency}
\title{Frequency App}
\usage{
app_frequency(df)
}
\arguments{
\item{df}{data}
}
\value{
No return value. This function is called for the side effect of
launching a shiny application.
}
\description{
Frequency App
}
\examples{
if (interactive()) {
app_frequency(mtcars)
}
}
