% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.R
\name{plot_box}
\alias{plot_box}
\title{Plot - Box}
\usage{
plot_box(
  data,
  x,
  y,
  theme,
  box_width,
  fill,
  outline,
  row,
  column,
  width,
  height,
  title,
  subtitle,
  xlab,
  ylab,
  caption,
  show_code,
  code
)
}
\arguments{
\item{data}{data}

\item{x}{variable on x axis}

\item{y}{variable on y axis}

\item{theme}{theme of plot}

\item{box_width}{width of box}

\item{fill}{fill}

\item{outline}{outline}

\item{row}{facet row in grid}

\item{column}{facet column in grid}

\item{width}{width of plot}

\item{height}{height of plot}

\item{title}{title of plot}

\item{subtitle}{subtitle of plot}

\item{xlab}{x-axis}

\item{ylab}{y-axis label}

\item{caption}{caption}

\item{show_code}{show the generated code TRUE or FALSE, (default TRUE)}

\item{code}{add custom code}
}
\value{
No return value. This function is called for the side effect of
launching a shiny application.
}
\description{
Plot - Box
}
\examples{
if (interactive()) {
  plot_box(mtcars)
}
}
