% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Plot - Boxplot}
\usage{
plot_boxplot(data)
}
\arguments{
\item{data}{data}
}
\value{
No return value. This function is called for the side effect of
launching a shiny application.
}
\description{
Plot - Boxplot
}
\examples{
if (interactive()) {
  plot_boxplot(mtcars)
}
}
