% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogram.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Plot - Histogram}
\usage{
plot_histogram(
  data,
  xvar,
  fill,
  outline,
  row,
  column,
  width,
  height,
  title,
  subtitle,
  xlab,
  ylab,
  caption,
  theme,
  code,
  show_code
)
}
\arguments{
\item{data}{data}

\item{xvar}{xvar}

\item{fill}{fill}

\item{outline}{outline}

\item{row}{facet row in grid}

\item{column}{facet column in grid}

\item{width}{width of plot}

\item{height}{height of plot}

\item{title}{title of plot}

\item{subtitle}{subtitle of plot}

\item{xlab}{x-axis}

\item{ylab}{y-axis label}

\item{caption}{caption}

\item{theme}{theme}

\item{code}{additional code}

\item{show_code}{show/hide code}
}
\value{
No return value. This function is called for the side effect of
launching a shiny application.
}
\description{
Plot - Histogram
}
\examples{
if (interactive()) {
  plot_histogram(mtcars)
}
}
