% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_line.R
\name{plot_line}
\alias{plot_line}
\title{Plot - Line}
\usage{
plot_line(
  data,
  xvar,
  yvar,
  theme,
  group,
  attr,
  code,
  row,
  column,
  width,
  height,
  title,
  subtitle,
  xlab,
  ylab,
  caption,
  show_code
)
}
\arguments{
\item{data}{data}

\item{xvar}{xvar}

\item{yvar}{yvar}

\item{theme}{theme}

\item{group}{group}

\item{attr}{attr}

\item{code}{additional code}

\item{row}{facet row in grid}

\item{column}{facet column in grid}

\item{width}{width of plot}

\item{height}{height of plot}

\item{title}{title of plot}

\item{subtitle}{subtitle of plot}

\item{xlab}{x-axis}

\item{ylab}{y-axis label}

\item{caption}{caption}

\item{show_code}{show/hide code}
}
\value{
No return value. This function is called for the side effect of
launching a shiny application.
}
\description{
Plot - Line
}
\examples{
if (interactive()) {
  plot_line(mtcars)
}
}
