% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot - Scatter}
\usage{
plot_scatter(data)
}
\arguments{
\item{data}{data}
}
\value{
No return value. This function is called for the side effect of
launching a shiny application.
}
\description{
Plot - Scatter
}
\examples{
if (interactive()) {
  plot_scatter(mtcars)
}
}
