% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_correlation.R
\name{stat_correlation}
\alias{stat_correlation}
\title{Stat - Correlation}
\usage{
stat_correlation(data, xvar, code, show_code)
}
\arguments{
\item{data}{data}

\item{xvar}{list of variables}

\item{code}{additional code}

\item{show_code}{show/hide code}
}
\value{
No return value. This function is called for the side effect of
launching a shiny application.
}
\description{
Stat - Correlation
}
\examples{
if (interactive()) {
stat_frequency(mtcars)
}
}
