% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_regression_linear.R
\name{stat_regression_linear}
\alias{stat_regression_linear}
\title{Stat - linear regression}
\usage{
stat_regression_linear(data, xvar, yvar, code, show_code)
}
\arguments{
\item{data}{data}

\item{xvar}{xvar}

\item{yvar}{yvar}

\item{code}{additional code}

\item{show_code}{show/hide code}
}
\value{
No return value. This function is called for the side effect of
launching a shiny application.
}
\description{
Stat - linear regression
}
\examples{
if (interactive()) {
  stat_regression_linear(mtcars)
}
}
