% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_ttest.R
\name{stat_ttest}
\alias{stat_ttest}
\title{Stat - T-test}
\usage{
stat_ttest(data, xvar, yvar, group, test, show_code)
}
\arguments{
\item{data}{data}

\item{xvar}{xvar}

\item{yvar}{yvar}

\item{group}{groups}

\item{test}{type of t-test}

\item{show_code}{show/hide code}
}
\value{
No return value. This function is called for the side effect of
launching a shiny application.
}
\description{
Stat - T-test
}
\examples{
if (interactive()) {
  stat_ttest(mtcars)
}
}
