% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot - Scatter}
\usage{
plot_scatter(
  data,
  xvar,
  yvar,
  color,
  shape,
  size,
  row,
  column,
  title,
  subtitle,
  caption,
  xlab,
  ylab,
  theme
)
}
\arguments{
\item{data}{data}

\item{xvar}{x}

\item{yvar}{y}

\item{color}{color}

\item{shape}{shape}

\item{size}{size}

\item{row}{row}

\item{column}{column}

\item{title}{title}

\item{subtitle}{subtitle}

\item{caption}{caption}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{theme}{theme}
}
\value{
No return value. This function is called for the side effect of
launching a shiny application.
}
\description{
Plot - Scatter
}
\examples{
if (interactive()) {
  plot_scatter(mtcars)
}
}
