% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_frequency.R
\name{stat_frequency}
\alias{stat_frequency}
\title{Frequency App}
\usage{
stat_frequency(data, xvar)
}
\arguments{
\item{data}{data}
}
\value{
No return value. This function is called for the side effect of
launching a shiny application.
}
\description{
Frequency App
}
\examples{
if (interactive()) {
stat_frequency(mtcars)
}
}
