% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_frequency.R
\name{stat_frequency}
\alias{stat_frequency}
\title{Frequency}
\usage{
stat_frequency(data, xvar)
}
\arguments{
\item{data}{data}

\item{xvar}{variable}
}
\value{
No return value. This function is called for the side effect of
launching a shiny application.
}
\description{
Frequency
}
\examples{
if (interactive()) {
stat_frequency(mtcars)
}
}
